+++ frontend/src/App.css
@@ -36,3 +36,36 @@
     transform: rotate(360deg);
   }
 }
+
+
+.nav{display:flex;gap:.75rem;padding:.75rem 1rem;border-bottom:1px solid #e5e7eb;}
+.nav .grow{display:flex;gap:.5rem;flex:1}
+.nav a{text-decoration:none;color:#374151;padding:.25rem .5rem}
+.nav .active{background:#eef2ff;border-radius:.375rem}
+
+.center{max-width:420px;margin:2rem auto;}
+.panel{display:grid;gap:.5rem}
+.panel input,.btn,button{padding:.6rem .75rem;border:1px solid #e5e7eb;border-radius:.5rem}
+
+.grid.hub{display:grid;gap:.75rem;grid-template-columns:repeat(2,minmax(0,1fr));max-width:520px;margin:2rem auto}
+.tile{display:flex;align-items:center;justify-content:center;height:56px;border:1px solid #e5e7eb;border-radius:.5rem;text-decoration:none;color:#111827}
+
+.profile{display:grid;gap:.5rem;place-items:center;margin:2rem auto;max-width:360px}
+.avatar{width:96px;height:96px;border-radius:9999px;background:#e5e7eb}
+.pill{padding:.5rem .75rem;background:#f3f4f6;border-radius:.5rem;width:100%;text-align:center}
+
+.wallpaper{height:80px;display:flex;align-items:center;padding:0 1rem;background:#f3f4f6;border-bottom:1px solid #e5e7eb}
+.toolbar{display:flex;gap:.5rem;padding:.5rem 1rem;border-bottom:1px solid #e5e7eb}
+.layout{display:grid;grid-template-columns:320px 1fr;gap:1rem;padding:1rem}
+.filters{display:grid;gap:.5rem}
+.cards{display:grid;gap:1rem;grid-template-columns:repeat(auto-fill,minmax(180px,1fr));padding:1rem}
+.card{border:1px solid #e5e7eb;border-radius:.5rem;height:120px;background:#fff}
+.card.tall{height:220px}
+.stack{display:grid;gap:.75rem}
+
+.cal{border:1px solid #e5e7eb;border-radius:.5rem}
+.cal-head{padding:.5rem;border-bottom:1px solid #e5e7eb}
+.cal-grid{display:grid;grid-template-columns:repeat(7,1fr)}
+.cal-dow{font-size:.8rem;padding:.5rem;background:#fafafa;border-bottom:1px solid #eee}
+.cal-cell{padding:.75rem .5rem;text-align:left;border:1px solid #f0f0f0;background:#fff}
+.cal-cell.muted{color:#9ca3af;background:#fafafa}
