+++ frontend/src/pages/About.jsx
@@ -1,0 +1,3 @@
+// About.jsx
+import React from "react";
+export default function About(){ return <main className="center"><h1>about us</h1></main>; }
