+++ frontend/src/pages/Explore.jsx
@@ -1,0 +1,20 @@
+import React from "react";
+
+export default function Explore() {
+  return (
+    <main className="explore">
+      <header className="wallpaper">[wallpaper of culture selected]</header>
+      <section className="cards">
+        <div className="card" />
+        <div className="card" />
+        <div className="card" />
+      </section>
+      <h3>new events coming (based on festivals)</h3>
+      <section className="cards">
+        <div className="card" />
+        <div className="card" />
+        <div className="card" />
+      </section>
+    </main>
+  );
+}
