+++ frontend/src/pages/Info.jsx
@@ -1,0 +1,3 @@
+// Info.jsx
+import React from "react";
+export default function Info(){ return <main className="center"><h1>info</h1></main>; }
