+++ frontend/src/pages/Profile.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+export default function Profile() {
+  return (
+    <main className="profile">
+      <div className="avatar" />
+      <div className="pill">name</div>
+      <div className="pill">age</div>
+      <div className="pill">country</div>
+      <div className="pill">culture</div>
+    </main>
+  );
+}
